+++ src/main/java/sshj/sshj/service/ClubService.java
@@ -14,7 +14,7 @@ public interface ClubService {
     void deleteClubNotice(long noticeId) throws Exception;
     void insertClubNoticeLike(long userId,long noticeId) throws Exception;
     void deleteClubNoticeLike(long userId,long noticeId) throws Exception;
-    int selectClubNoticeCnt(long noticeId) throws Exception;
+    long selectClubNoticeCnt(long noticeId) throws Exception;
     void insertClubSubs(long userId, long clubId) throws Exception;
     void deleteClubSubs(long userId, long clubId) throws Exception;
